+++ starport/services/scaffolder/module.go
@@ -3,6 +3,7 @@ package scaffolder
 import (
 	"context"
 	"errors"
+	"fmt"
 	"go/parser"
 	"go/token"
 	"os"
@@ -20,8 +21,56 @@ import (
 const (
 	wasmImport        = "github.com/CosmWasm/wasmd"
 	apppkg            = "app"
+	moduleDir         = "x"
 	wasmVersionCommit = "b30902fe1fbe5237763775950f729b90bf34d53f"
 )
+
+// CreateModule creates a new empty module in the scaffolded app
+func (s *Scaffolder) CreateModule(moduleName string) error {
+	version, err := s.version()
+	if err != nil {
+		return err
+	}
+	// Check if the module already exist
+	ok, err := ModuleExists(s.path, moduleName)
+	if err != nil {
+		return err
+	}
+	if ok {
+		return errors.New(fmt.Sprintf("The module %v already exists.", moduleName))
+	}
+	path, err := gomodulepath.ParseFile(s.path)
+	if err != nil {
+		return err
+	}
+
+	var (
+		g    *genny.Generator
+		opts = &module.CreateOptions{
+			ModuleName: moduleName,
+			ModulePath: path.RawPath,
+			AppName:    path.Package,
+		}
+	)
+	if version == cosmosver.Launchpad {
+		g, err = module.NewCreateLaunchpad(opts)
+	} else {
+		g, err = module.NewCreateStargate(opts)
+	}
+	if err != nil {
+		return err
+	}
+	run := genny.WetRunner(context.Background())
+	run.With(g)
+	if err := run.Run(); err != nil {
+		return err
+	}
+	pwd, err := os.Getwd()
+	if err != nil {
+		return err
+	}
+	return s.protoc(pwd, version)
+}
 
 // ImportModule imports specified module with name to the scaffolded app.
 func (s *Scaffolder) ImportModule(name string) error {
@@ -50,7 +99,7 @@ func (s *Scaffolder) ImportModule(name string) error {
 	if err != nil {
 		return err
 	}
-	g, err := module.NewImport(&module.Options{
+	g, err := module.NewImport(&module.ImportOptions{
 		Feature: name,
 		AppName: path.Package,
 	})
@@ -60,6 +109,24 @@ func (s *Scaffolder) ImportModule(name string) error {
 	run := genny.WetRunner(context.Background())
 	run.With(g)
 	return run.Run()
+}
+
+func ModuleExists(appPath string, moduleName string) (bool, error) {
+	abspath, err := filepath.Abs(filepath.Join(appPath, moduleDir, moduleName))
+	if err != nil {
+		return false, err
+	}
+
+	_, err = os.Stat(abspath)
+	if err == nil {
+		// The module already exists
+		return true, nil
+	}
+	if os.IsNotExist(err) {
+		return false, nil
+	}
+	// Error reading the directory
+	return false, err
 }
 
 func isWasmImported(appPath string) (bool, error) {